streamlit/requirements.txt
streamlit
torch
Pillow
opencv-python-headless
numpy
matplotlib
